refactor: adapter 패키지 분리
